New UnknownPointSet for runtime known points
Allows a vector of points which are unknown at compile time to be
expressed at runtime. The expression given to the UnknownPointSet will
be runtime tabulated if it has a name that begins with 'rt_'. Points
are a non-indexable/non-iterable UnknownPointsArray object that mocks
the usual points attribute of a point set.

Also tidy some other comments.