Avoid tuple hash collisions
On python versions prior to 3.8 use of the inbuilt hash function can
easily lead to hash collisions e.g.

hash( (0.5, 0.0, 0.5) ) == hash( (0.5, 0.5, 0.0) )

In fact it turns out that the use of the in-build hash function is not
recommended for this reason.

Dictionaries have other checks in built so I now how the Q Dictionary of
Keys with a flattened tuple directly instead of wrapping it in the built
in hash function.